% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{addGridIfTrue}
\alias{addGridIfTrue}
\title{add a grid if the first argument is TRUE}
\usage{
addGridIfTrue(plot.grid, xPositions, yPositions = graphics::axTicks(2),
  col = "grey", lty = 3, ...)
}
\arguments{
\item{plot.grid}{logical. if TRUE the grid is plotted, else not.}

\item{xPositions}{x positions of the vertical grid lines}

\item{yPositions}{y positions of the horizontal grid lines}

\item{\dots}{additional arguments passed to abline}
}
\description{
# add a grid if the first argument is TRUE
}
