% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{apply_elements_text}
\alias{apply_elements_text}
\title{Apply themes with given text properties to a plot}
\usage{
apply_elements_text(x, elements)
}
\arguments{
\item{x}{ggplot object}

\item{elements}{objects as created by function 
\code{\link[ggplot2]{element_text}}}
}
\description{
Apply themes with given text properties to a plot
}
