% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{appropriateLimits}
\alias{appropriateLimits}
\title{default limits for plotting values}
\usage{
appropriateLimits(x, limits = NULL, default = c(0, 1))
}
\arguments{
\item{x}{vector of numeric or POSIXt (min, max must be able to be applied).}

\item{limits}{vector of two elements. Default: NULL.}

\item{default}{vector of two elements. Default: c(0, 1)}
}
\value{
returns \emph{limits} if \emph{limits} is a vector of two non-NA values. 
  If the first element of \emph{limits} is NA it is replaced with the 
  minimum of \emph{x} (or with \emph{default[1]} if the minimum is NA). If
  the second element of \emph{limits} is NA it is replaced with the maximum
  of \emph{x} (or with \emph{default[2]} if the maximum is NA).
}
\description{
default limits for plotting values
}
