% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{args_mybarplot}
\alias{args_mybarplot}
\title{Create argument list for mybarplot}
\usage{
args_mybarplot(cex.axis = graphics::par("cex.axis"),
  cex.names = graphics::par("cex.axis"), ..., args.text = NULL,
  args.yaxis = NULL)
}
\arguments{
\item{cex.axis}{character expansion factor of axis labels}

\item{cex.names}{character expansion factor of bar group labels}

\item{\dots}{additional arguments to barplot}

\item{args.text}{list of arguments that are given to \code{text} being called to plot the
bar group labels. Use \code{\link{args_text}} to create such a list.}

\item{args.yaxis}{list of arguments that are given to \code{axis} being called to plot the
y axis. Use \code{\link{args_yaxis}} to create such a list.}
}
\description{
Create argument list for mybarplot
}
