% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{arrange_in_pdf}
\alias{arrange_in_pdf}
\title{Arrange Plots in PDF file (DIN A4)}
\usage{
arrange_in_pdf(plots, landscape = TRUE, ...)
}
\arguments{
\item{plots}{list of ggplot objects}

\item{landscape}{passed to \code{\link[kwb.utils]{preparePdf}}}

\item{\dots}{arguments passed to \code{\link[gridExtra:arrangeGrob]{grid.arrange}}}
}
\description{
Arrange ggplot objects with gridExtra::grid.arrange and write the result
  to a PDF file in DIN A4 format
}
