% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{arrange_in_png}
\alias{arrange_in_png}
\title{Arrange Plots in PNG file}
\usage{
arrange_in_png(plots, filename, ...)
}
\arguments{
\item{plots}{list of ggplot objects}

\item{filename}{full path to the PNG file to be written. The extension ".png" will be 
automatically appended if it is omitted}

\item{\dots}{arguments passed to \code{\link[gridExtra]{grid.arrange}}}
}
\description{
Arrange ggplot objects with gridExtra::grid.arrange and write the result
  to a PNG file
}
