% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bestRowColumnSetting}
\alias{bestRowColumnSetting}
\title{best nrow/ncol setting for n plots}
\usage{
bestRowColumnSetting(n, target.ratio = 1)
}
\arguments{
\item{n}{number of plots to be placed in a matrix of equally sized plot cells}

\item{target.ratio}{desired height/width ratio within each plot (ignoring margins). Default: 1}
}
\value{
named vector of two elements with the first element (\code{nrow})
  representing the number of rows and the second element (\code{ncol})
  representing the number of columns for an optimal plot grid to be used for
  \code{n} plots in the current plot region
}
\description{
Number of rows and columns in an optimal plot grid for n plots
}
\examples{
  # save current graphical parameter setting
  old.par <- graphics::par(no.readonly = TRUE)
  
  for (i in 2:5) {
    
    graphics::par(mfrow = kwb.plot::bestRowColumnSetting(i))
    
    for (j in 1:i) {
      plot(seq_len(j), main = paste0("j = ", j, "/", i))
    }      
  }
  
  # restore graphical parameter setting
  graphics::par(old.par)  
  

}
