% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{cmToUserWidthAndHeight}
\alias{cmToUserWidthAndHeight}
\title{Convert Length in Centimetres to Lengths in User Coordinates}
\usage{
cmToUserWidthAndHeight(cm)
}
\arguments{
\item{cm}{length in centimetres}
}
\description{
Convert Length in Centimetres to Lengths in User Coordinates
}
