% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo.adj}
\alias{demo.adj}
\title{Demo of Graphical Parameter "adj"}
\usage{
demo.adj(text = "Text", srt = c(0, 90), cex = 1, ..., to.pdf = FALSE)
}
\arguments{
\item{text}{text to be plotted}

\item{srt}{The string rotation in degrees (see \code{par})}

\item{cex}{character expansion factor}

\item{\dots}{further arguments passed to \code{text}}

\item{to.pdf}{if \code{TRUE} the output goes into a PDF file}
}
\description{
Demonstration of the effect of the graphical parameter "adj" (horizontal and
  vertical adjustment)
}
\examples{
  kwb.plot::demo.adj()
  kwb.plot::demo.adj("Exampletext", srt = 30)
  kwb.plot::demo.adj(srt = c(0, 45, 90), to.pdf = TRUE)  
  
}
