% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlevel.R
\name{.defaultPlotParameter}
\alias{.defaultPlotParameter}
\title{Default Plot Parameters for Hydraulic Variables}
\usage{
.defaultPlotParameter(parameterName, variableName)
}
\arguments{
\item{parameterName}{one of c("type", "col", "ylab", "pch")}

\item{variableName}{variable name matching
"(Q|H|v)[.](raw|signal|interpol|pred)"}
}
\description{
Default Plot Parameters for Hydraulic Variables
}
