% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{drawBoxplot}
\alias{drawBoxplot}
\title{Draw Boxplot Icon}
\usage{
drawBoxplot(centre.x, centre.y, boxwidth.cm = 1, boxheight.cm = boxwidth.cm,
  whisker.cm = boxheight.cm)
}
\arguments{
\item{centre.x}{x coordinate in user coordinates around which the box is to be drawn}

\item{centre.y}{y coordinate in user coordinates around which the box is to be drawn}

\item{boxwidth.cm}{width of the box in cm. Default: 1}

\item{boxheight.cm}{height of the box in cm. Default: \emph{boxwidth.cm}}

\item{whisker.cm}{length of the whiskers in cm. Default: \emph{boxheight.cm}}
}
\description{
draws a symmetric boxplot icon around a centre
}
\examples{
  
  ### prepare a simple plot area
  plot(1:5)
  
  ### draw a box around the centre at (2, 2) with default proportions
  drawBoxplot(2, 2, boxwidth.cm = 1)
  
  ### draw a box around the centre at (3, 3) with differing width and height
  drawBoxplot(3, 3, boxwidth.cm = 2, boxheight.cm = 1)
  
  ### draw a box around the centre at (4, 4) with modified whisker lengths
  drawBoxplot(4, 4, boxwidth.cm = 0.5, boxheight.cm = 1.5, whisker.cm = 0.5)
  
}
