% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{element_types}
\alias{element_types}
\title{element types}
\usage{
element_types(secondary = FALSE)
}
\arguments{
\item{secondary}{if \code{TRUE} all parameters are returned otherwise those
parameters related to the x axis on top, the y axis on the right or the 
distict x an y parameters related to the "stip.text" parameter are omitted}
}
\description{
element types
}
