% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{getCurrentLimits}
\alias{getCurrentLimits}
\title{Get Current xlim and ylim}
\usage{
getCurrentLimits(
  type = "xy",
  xaxs = graphics::par("xaxs"),
  yaxs = graphics::par("yaxs")
)
}
\arguments{
\item{type}{one of "x", "y" or "xy", see section "Value"}

\item{xaxs}{value of argument \emph{xaxs} when calling a plot function last, see
\code{par}}

\item{yaxs}{value of argument \emph{yaxs} when calling a plot function last, see
\code{par}}
}
\value{
\emph{type} = "x" or "y": vector of two elements corresponding to xlim or
  ylim, respectively. \emph{type} = "xy": list with elements \emph{xlim} and
  \emph{ylim} each of which is a vector of two elements.
}
\description{
get the values of xlim and/or ylim that created the last plot
}
