% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getPlotCharacterConstants}
\alias{getPlotCharacterConstants}
\title{selected, named pch constants}
\usage{
getPlotCharacterConstants()
}
\value{
list of plot character constants with each element being named according
  to the appearence of the plot character, e.g. "CIRCLE", "TRIANGLE", ...
}
\description{
selected, named plot character (pch) constants
}
