% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{getPlotRegionRatio}
\alias{getPlotRegionRatio}
\title{Current plot region's ratio}
\usage{
getPlotRegionRatio()
}
\value{
numeric of length one representing the width/height ratio of the current
  plot region
}
\description{
Current plot region's height/width ratio
}
