% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{getPlotRegionSizeInUserCoords}
\alias{getPlotRegionSizeInUserCoords}
\title{Size of Current Plot Region in User Coordinates}
\usage{
getPlotRegionSizeInUserCoords()
}
\value{
list with elements \emph{width}, \emph{height}, \emph{left}, \emph{bottom}, 
  \emph{right}, \emph{top}
}
\description{
Size of Current Plot Region in User Coordinates
}
