% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_text_lines.R
\name{ggplot_text_lines}
\alias{ggplot_text_lines}
\title{Plot Text Only}
\usage{
ggplot_text_lines(
  x,
  max_rows = max(10, length(x)),
  size = 3,
  max_chars = 170,
  margins_cm = c(0, 0, 0, 0)
)
}
\arguments{
\item{x}{vector of character representing the rows of text to be plotted from
top to bottom}

\item{max_rows}{number of rows to be prepared in the plot area. 
Increase/decrease this number to decrease/increase the line spacing}

\item{size}{text size. Default: 3}

\item{max_chars}{maximum number of characters to be used in each row. 
Longer lines that shortened to this number using 
\code{\link[kwb.utils]{shorten}}}

\item{margins_cm}{vector of four numerics giving the bottom, left, top and
top margins in cm, respectively. Default: \code{c(0, 0, 0, 0)}}
}
\description{
Plot Text Only
}
\examples{
ggplot_text_lines(paste(c("first", "second", "third"), "row"))

}
