% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{inchesToUserWidthAndHeight}
\alias{inchesToUserWidthAndHeight}
\title{Convert Length in Inches to Lengths in User Coordinates}
\usage{
inchesToUserWidthAndHeight(inches)
}
\arguments{
\item{inches}{length in inches}
}
\description{
Convert Length in Inches to Lengths in User Coordinates
}
