% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{inLimits}
\alias{inLimits}
\title{are values within limits (e.g. xlim, ylim)?}
\usage{
inLimits(x, limits)
}
\arguments{
\item{x}{vector of values}

\item{limits}{two-element vector of (lower and upper) limits}
}
\value{
vetor of logical. Each element corresponds to an element in \emph{x} and
  is TRUE if the element is within the limits or FALSE if the element is
  out of the limits.
}
\description{
are values within limits (e.g. xlim, ylim)?
}
