% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{labelPositionY}
\alias{labelPositionY}
\title{y-positions for labels}
\usage{
labelPositionY(n, y0 = 0, bandheight = 0.1, group.size = 3,
  alternating = TRUE)
}
\arguments{
\item{n}{number of y-positions to be generated}

\item{y0}{y-positions around which the generated values will "alternate"}

\item{bandheight}{height of band "around" (alternating == FALSE) or above (alternating ==
TRUE) y0 as a fraction of the plot region height (e.g. 0.1 for 10
percent). Default: 0.1}

\item{group.size}{number of labels that are to be placed in on "group" in
which label positions are modified along the y axis}

\item{alternating}{if \code{TRUE} (default) the label positions are 
alternating between positive and negative values}
}
\description{
alternating y-positions to be used for label placement
}
