% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{niceLabels}
\alias{niceLabels}
\title{Nice Labels}
\usage{
niceLabels(label, labelstep = NULL, labelpos = NULL, mindist = 1,
  offset = 0)
}
\arguments{
\item{label}{vector of labels}

\item{labelstep}{step width in which labels are to be printed, e.g. labelStep = 2 is used
for plotting every second label.}

\item{labelpos}{label positions}

\item{mindist}{minimum distance between labels}

\item{offset}{offset by which the index of the first remaining label (default: 1) is 
shifted. Default: 0, i.e. the labels at indices \emph{1}, \emph{1 + 
1*labelstep}, \emph{1 + 2*labelstep}, etc. remain. Offset = 1: the labels 
at indices \emph{2}, \emph{2 + 1*labelstep}, \emph{2 + 2*labelstep}, etc.
remain.}
}
\description{
Generates a nice vector of labels by suppressing labels at certain
  positions. If a \emph{labelStep} is given, only every \emph{labelStep}-th 
  label is considered. If a vector \emph{labpos} of label positions and
  a minimum distance \emph{mindist} of labels is given, it is guaranteed
  that the distance between labels to be shown is at least \emph{mindist}.
}
\examples{
  
  x <- matrix(1:12, nrow = 3)
  
  names.arg <- rep(1:4, each = 3)
  
  mybarplot <- function(x, ...) {
    barplot(x, horiz = TRUE, las = 1, beside = TRUE, ...)
  }
    
  mybarplot(x, names.arg = names.arg)
  mybarplot(x, names.arg = niceLabels(names.arg, labelstep = 3))
  mybarplot(x, names.arg = niceLabels(names.arg, labelstep = 3, offset = 1))
  mybarplot(x, names.arg = niceLabels(names.arg, labelstep = 3, offset = 2))
  
}
