% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{output_to_png}
\alias{output_to_png}
\title{Redirect Output of Plot Function to PNG file}
\usage{
output_to_png(FUN, args, filename, size = unlist(kwb.utils::DIN.A4()),
  units = "cm", res = 300, ..., dbg = TRUE)
}
\arguments{
\item{FUN}{plot function to be called}

\item{args}{list of arguments given to the plot function \code{FUN}}

\item{filename}{full path to the PNG file to be written. The extension ".png" will be 
automatically appended if it is omitted}

\item{size}{vector containing the width (first element) and height (second element)
of the plot, as passed to \code{\link[grDevices]{png}}. By default 
the dimensions of a DIN A4 page are used.}

\item{units}{units as passed to \code{\link[grDevices]{png}}}

\item{res}{resolution as passed to \code{\link[grDevices]{png}}}

\item{\dots}{further arguments passed to \code{\link[grDevices]{png}}}

\item{dbg}{if \code{TRUE} a message "Plotting to <filename>..." is shown}
}
\description{
Redirect the output of a plot function to a PNG file
}
