% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{plot_all_to_pdf}
\alias{plot_all_to_pdf}
\title{Print all ggplot Objects to a PDF file}
\usage{
plot_all_to_pdf(plots, landscape = TRUE, ...)
}
\arguments{
\item{plots}{list of ggplot objects}

\item{landscape}{passed to \code{\link[kwb.utils]{preparePdf}}}

\item{\dots}{further arguments passed to \code{\link[kwb.utils]{preparePdf}}}
}
\description{
Print all ggplot objects to a PDF file
}
