% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_areas.R
\name{plot_curve_areas_gg}
\alias{plot_curve_areas_gg}
\title{Plot Filled Areas below Curve Lines using ggplot}
\usage{
plot_curve_areas_gg(x = seq_along(y_list[[1]]), y_list, col = NULL,
  stack = FALSE, legend = TRUE, line_colour = "black")
}
\arguments{
\item{x}{x positions of each curve given in \code{y_list}}

\item{y_list}{list of vectors of y positions each of which must be as long as
\code{x}}

\item{col}{vector of the same length as \code{y_list} giving the colours of
the areas to be filled}

\item{stack}{if \code{TRUE} (the default if \code{FALSE}) the areas are
stacked on top of each other otherwise they are overlayed.}

\item{legend}{if \code{TRUE} (default) the standard legend is shown, else it 
is hidden}

\item{line_colour}{colour of the curve lines}
}
\description{
Plot Filled Areas below Curve Lines using ggplot
}
