% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlevel.R
\name{plot_variable}
\alias{plot_variable}
\title{plot_variable}
\usage{
plot_variable(hydraulicData, variableName = names(hydraulicData)[[2]],
  type = .defaultPlotParameter("type", variableName),
  col = .defaultPlotParameter("col", variableName),
  ylab = .defaultPlotParameter("ylab", variableName),
  pch = .defaultPlotParameter("pch", variableName), xlim = NULL,
  ylim = NULL, add = FALSE, plot.grid = TRUE, innerMargins = c(0, 0, 0,
  0), absolute = FALSE, reverse.ylim = FALSE, add.time.axis = TRUE,
  time.axis.in.margins = FALSE, ...)
}
\arguments{
\item{hydraulicData}{data frame with at least two columns. First column is expected to contain
the timestamps}

\item{variableName}{default: name of second column}

\item{absolute}{if TRUE, innerMargins are to be interpreted as absolute values instead of 
fractions of the plot region. Default: FALSE}

\item{reverse.ylim}{if TRUE (default), the y axis is reversed, i.e. the bars are heading from
top to bottom}
}
\description{
plot_variable
}
