% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_level_frequencies.R
\name{plotLevelFrequencies}
\alias{plotLevelFrequencies}
\title{Plot Frequencies of Level Combinations}
\usage{
plotLevelFrequencies(data, vars_1 = NULL, vars_2 = NULL, relative = TRUE)
}
\arguments{
\item{data}{data frame}

\item{vars_1}{first vector of column names}

\item{vars_2}{second vector of column names}

\item{relative}{if \code{TRUE} (default)}
}
\value{
list of objects of class "ggplot"
}
\description{
For combinations of columns \code{c1} and \code{c2} in \code{data},
ggplot-objects are generated each of which shows the number (\code{relative =
FALSE}) or percentage (\code{relative = TRUE}) of the different possible
combinations of values in \code{c1}, and \code{c2}. The combinations of
column names are created from the permutation of column names given in
\code{vars_1} and \code{vars_2}, respectively.
}
\examples{
data <- data.frame(
  fruit = c("apple", "cherry", "apple", "banana", "cherry"),
  colour = c("green", "red", "red", "yellow", "green"),
  size = c("small", "small", "big", "small", "big")
)

plotLevelFrequencies(data)

}
