% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlevel.R
\name{plotRain}
\alias{plotRain}
\title{Plot Rain Data as Bars}
\usage{
plotRain(
  timestamps,
  values,
  gaugeName = "",
  xlim = NULL,
  ylim = NULL,
  signal.width = NA,
  shift.to.begin = 0,
  col = "blue",
  pch = 16,
  reverse.ylim = TRUE,
  las = 1,
  innerMargins = c(0, 0, 0, 0),
  add.time.axis = TRUE,
  ...
)
}
\arguments{
\item{timestamps}{timestamps}

\item{values}{precipitation values in mm}

\item{gaugeName}{Name of rain gauge to appear in the y label}

\item{xlim}{x limits}

\item{ylim}{y limits}

\item{signal.width}{width of time intervals represented by timestamps, in seconds. If NA 
(default) this value is deduced from the time differences in 
\emph{timestamps}}

\item{shift.to.begin}{value in seconds by which the timstamps need to be shifted in order to get
the begin of the time intervals that the timestamps represent. Default: 0
(meaning that the timestamps represent the beginnings of the time 
intervals). A value of \emph{signal.width} would mean that the timestamps
represent the ends of the time intervals and a value of 
\emph{signal.width/2} would mean that the timestamps represent the 
midpoints of the time intervals that they represent.}

\item{col}{fill colour of bars. Default: "blue"}

\item{pch}{plot character to be used for plotting data points (at the given
timestamps), additionally to plotting bars. Set to NA in order to suppress
plotting these points}

\item{reverse.ylim}{if TRUE (default), the y axis is reversed, i.e. the bars are heading from
top to bottom}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. See help for par.}

\item{innerMargins}{passed to \code{\link{plot_variable}}}

\item{add.time.axis}{passed to \code{\link{plot_variable}}}

\item{\dots}{further arguments passed to \code{\link{points}}}
}
\description{
Plot Rain Data as Bars
}
