% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{preview_themes}
\alias{preview_themes}
\title{Preview the Effects of Themes on a Plot}
\usage{
preview_themes(x = example_plot(), themes = ggplot_themes(),
  to_pdf = TRUE, landscape = TRUE, ...)
}
\description{
Preview the effects of ggplot2-themes on a given plot
}
