% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_labels.R
\name{set_titles}
\alias{set_titles}
\title{Set the Titles in a List of ggplot Objects}
\usage{
set_titles(plots, title, indices = seq_along(plots), ...)
}
\arguments{
\item{plots}{list of ggplot objects as returned by \code{ggplot}}

\item{title}{title (character) to be given to each plot or to the plots
selected by their \code{indices}}

\item{indices}{indices of the plots to which the title is to be given. By
default the title is given to all plots}

\item{\dots}{additional arguments to \code{\link{set_labels}}, such as 
\code{action}}
}
\description{
Set the Titles in a List of ggplot Objects
}
