% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_labels.R
\name{set_xlabs}
\alias{set_xlabs}
\title{Set the x Axis Label in a List of ggplot Objects}
\usage{
set_xlabs(plots, xlab, indices = seq_along(plots))
}
\arguments{
\item{plots}{list of ggplot objects as returned by \code{ggplot}}

\item{xlab}{x axis label (character) to be given to each plot or to the plots selected
by their \code{indices}}

\item{indices}{indices of the plots to which the x axis label is to be given. By default
the x axis label is given to all plots}
}
\description{
Set the x Axis Label in a List of ggplot Objects
}
