% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackplot.R
\name{stackplotLayout}
\alias{stackplotLayout}
\title{Create Layout for "stackplot"}
\usage{
stackplotLayout(heights.cm, margins.top.cm, margins.bottom.cm, show = FALSE)
}
\description{
Create Layout for "stackplot"
}
