% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{to_element_themes}
\alias{to_element_themes}
\title{to element themes}
\usage{
to_element_themes(names, element)
}
\arguments{
\item{names}{name of the argument given to \code{\link[ggplot2]{theme}}}

\item{element}{value of the argument given to \code{\link[ggplot2]{theme}}}
}
\description{
to element themes
}
