% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{userCoordinatesToLimits}
\alias{userCoordinatesToLimits}
\title{userCoordinatesToLimits}
\usage{
userCoordinatesToLimits(userCoordinates)
}
\value{
list with elements \emph{xlim} and \emph{ylim}, each of which is a numeric
  vector of length two.
}
\description{
back-calculate xlim and ylim from user coordinates of plot region
}
