% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{userWidthAndHeightToCm}
\alias{userWidthAndHeightToCm}
\title{Convert Lenghts in User Coordinates to Lengths in Centimetres}
\usage{
userWidthAndHeightToCm(width.user, height.user)
}
\arguments{
\item{width.user}{width in user coordinates}

\item{height.user}{height in user coordinates}
}
\description{
Convert Lenghts in User Coordinates to Lengths in Centimetres
}
