% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{addLabels}
\alias{addLabels}
\title{Add Labels}
\usage{
addLabels(x, labels = as.character(x), y0 = 0, bandheight = 0.1,
  col = "black", group.size = 3, alternating = FALSE,
  col.line = "black", lty = 1, lty.horiz.line = 0, adj = -0.1,
  cex = 0.7)
}
\arguments{
\item{bandheight}{height of band "around" (alternating == FALSE) or above (alternating ==
TRUE) y0 as a fraction of the plot region height (e.g. 0.1 for 10
percent). Default: 0.1}
}
\description{
add labels at given x-positions to plot (with alternating y positions
  to avoid overlapping labels)
}
