% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{args_text}
\alias{args_text}
\title{Create argument list for text}
\usage{
args_text(y.abs = NULL, y.rel = -0.01, srt = 90, adj = c(1, 0.5),
  ...)
}
\arguments{
\item{y.abs}{absolute y positions}

\item{y.rel}{relative y positions}

\item{srt}{see \code{\link{text}}}

\item{adj}{see \code{\link{text}}}

\item{\dots}{further arguments to \code{\link{text}}}
}
\value{
list of arguments that may be used as \code{args.text} in
  \code{\link{mybarplot}}
}
\description{
Create argument list for text
}
