% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{args_yaxis}
\alias{args_yaxis}
\title{Create argument list for axis}
\usage{
args_yaxis(side = 2, at = graphics::axTicks(side), labels = paste0(at,
  unit), unit = "", las = 1, ...)
}
\value{
list of arguments that may be used as \code{args.yasis} in
  \code{\link{mybarplot}}
}
\description{
Create argument list for axis
}
