% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getPlotCharacterConstants}
\alias{getPlotCharacterConstants}
\title{List of Named Constants for Graphical Parameter pch}
\usage{
getPlotCharacterConstants()
}
\value{
list of plot character constants with each element being named according
  to the appearence of the plot character, e.g. "CIRCLE", "TRIANGLE", ...
}
\description{
List of Named Constants for Graphical Parameter pch
}
\keyword{internal}
