% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlevel.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot a Variable of a Data Frame}
\usage{
plot_variable(
  hydraulicData,
  variableName = names(hydraulicData)[[2]],
  type = .defaultPlotParameter("type", variableName),
  col = .defaultPlotParameter("col", variableName),
  ylab = .defaultPlotParameter("ylab", variableName),
  pch = .defaultPlotParameter("pch", variableName),
  xlim = NULL,
  ylim = NULL,
  add = FALSE,
  plot.grid = TRUE,
  innerMargins = c(0, 0, 0, 0),
  absolute = FALSE,
  reverse.ylim = FALSE,
  add.time.axis = TRUE,
  time.axis.in.margins = FALSE,
  ...
)
}
\arguments{
\item{hydraulicData}{data frame with at least two columns. First column is expected to contain
the timestamps}

\item{variableName}{default: name of second column}

\item{type}{plot type, passed to \code{\link{points}}}

\item{col}{colour, passed to \code{\link{points}}}

\item{ylab}{y label, passed to \code{\link{plot}}}

\item{pch}{plot character, passed to \code{\link{points}}}

\item{xlim}{x limits, passed to \code{\link{plot}}}

\item{ylim}{y limits, passed to \code{\link{plot}}}

\item{add}{if \code{TRUE} points are added to an existing plot}

\item{plot.grid}{if \code{TRUE} a grid is drawn}

\item{innerMargins}{margins (bottom, left, top, right) within the plot area}

\item{absolute}{if \code{TRUE}, innerMargins are to be interpreted as absolute values instead of 
fractions of the plot region. Default: FALSE}

\item{reverse.ylim}{if \code{TRUE} (default), the y axis is reversed, i.e. the bars are heading from
top to bottom}

\item{add.time.axis}{if \code{TRUE} a time axis is drawn}

\item{time.axis.in.margins}{if \code{TRUE} the time axis is drawn only within
the plot area excluding the margins}

\item{\dots}{additional arguments passed to \code{\link{plot}}}
}
\description{
Plot a Variable of a Data Frame
}
