% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_labels.R
\name{set_labels}
\alias{set_labels}
\title{Set the Labels in a List of ggplot Objects}
\usage{
set_labels(plots, ..., indices = seq_along(plots))
}
\arguments{
\item{plots}{list of ggplot objects as returned by \code{ggplot}}

\item{\dots}{name-value pairs as given to \code{\link[ggplot2]{labs}}. Possible names
are e.g. "x", "title", "subtitle", "caption". The values are vectors
of character that are recycled to the length of \code{indices}. They are
used as labels given to each plot or to the plots selected by their
\code{indices}.}

\item{indices}{indices of the plots to which the label is to be given. By default the
label is given to all plots}
}
\description{
Set the Labels in a List of ggplot Objects
}
