% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackplot.R
\name{stackplot}
\alias{stackplot}
\title{Plots on Top of Each Other}
\usage{
stackplot(functionCalls, heights.cm = 5, margins.top.cm = 0,
  margins.bottom.cm = 0, margins.left.cm = 3, margins.right.cm = 1,
  envir = parent.frame())
}
\arguments{
\item{functionCalls}{list of expressions each of which is expected to create a plot when
being evaluated with eval}

\item{envir}{environment in which the expressions given in functionCalls are to be
evaluated. Default: parent.frame()}
}
\description{
Plots on Top of Each Other
}
